Version 0.9.8
Fixed the z-perspective-correctness in the Concourse and 2D menus.
(I did *not* have to divide by w). Also fixed the brackets (they are
now properly rendered *behind* the targetted ships) -- the bug was
in SBSVertexShader and the way output.pos.z was being written to.
The code here can probably be simplified even more. Changed the frustum
to make it shorter (didn't seem to have had any noticeable effect; but
using a lower value makes everything disappear).