Using W as linear Z instead of logarithmic sz
In this version, I started using W as Z coordinate instead of
using the logarithmic sz. Things got messed up again, because now
we need to specify a metric z_override *and* a logarithmic sz_override;
but it works. The skybox has already been fixed and placed at "infinity".
Other elements may also be fixed in the same way; but it's probably better
to recompute sz to allow for perspective-correct texturing.

Also added a generic rigid transform matrix to re-enable lean left/right,
boost up/down and zoom forward/backward (but this last one causes texturing
to break because it's no longer perspective-correct).