Linear Z + Perspective Correct Zoom; but broken engine glow.
In this version, I'm using W to reconstruct linear Z and using
a logarithmic formula for sz. This fixed the zoom so now it's
perspective-correct; but the engine glow does not have the correct
depth (it shows through the geometry). I also added support for roll
and re-enabled the keys to lean right/left, boost up/down and lean
forward/backward.

Several settings have to be specified in meters now:

HUD_parallax = 0.970
GUI_parallax = 0.495
GUI_target_parallax = -0.025
Text_parallax = 0.45

To prevent the roll from showing, I also increased the scale a little
bit:

3d_window_size = 1.80