Probable fix for positional tracking by inverting the rotation matrix.
This version enables limits for the positional tracking, and uses the
rotation matrix inverse to compensate for pitch/yaw already being applied
by XWA in the Hook. Also renamed all of the "parallax" parameters in
vrparams.cfg to "depth" or "relative_depth" and all of them are in meters
now. Minor code cleanup. Disabled 3D capture code again.