HUD is now static even when leaning, dynamic targeting comp
The aiming HUD now stays "in place" even when the user is leaning.
Before this change, the HUD would move around when the user is leaning
because it was "stuck" to the glass in the cockpit. This seemed sensible;
but made it hard to use it for aiming because it's no longer aligned with
the user's eye. With this fix, the HUD follows the user and it can still
be used for aiming.

Added code to load and replace textures at run-time. The cockpit targeting
computer at low res can now be replaced.

Added a new renderTargetView that is used to render the currently-selected
craft. This renderTargetView is associated with a new offscreenBuffer that
can be used to dynamically replace the targeting computer texture. In other
words, the targeted craft now appears in the cockpit. This is the first
version of this dynamic targeting computer and still needs a lot of polishing;
but it works!