Dynamic Cockpit Targeting Computer. Also fixed double-free bug.
DeleteStereoVertices was doing a double-free. Fixed by setting the pointers
to NULL after deleting them. The project has been upgraded to compile in debug
mode with DirectTK (DirectTK must also be compiled in Multithreaded Debug mode;
*not* Multithreaded Debug DLL).

The projection matrix for the dynamic targeting computer must be fixed in the
SteamVR version (I'm not sure what to use).