Added more vrparams, fixed non-VR mode.
Non-VR mode is now fully-functional. Added the following vrparams:

floating_aiming_HUD
dynamic_cockpit_enabled
concourse_animations_at_25fps

I still see some crashes when the dynamic cockpit is enabled. The new
texture is now set to NULL after it's released; but I'm not sure that
fixed the problem. Also, when the targeted craft is destroyed, it's still
displayed in the dynamic cockpit.