Removed additional vertex arrays. Added FreePIE support.
Removed the g_OrigVerts and g_3DVerts arrays. All vertex processing
is done in the GPU now and g_OrigVerts is now only a pointer that
is a convenience to access the command buffer.

Added FreePIE support for 6dof in DirectSBS mode. I can't test this
very well now, so I'm not sure it's 100% right. I had to change some
signs too because DirectSBS is now using the same coord system as
SteamVR. Will need to re-check the signs later.

Added extra vrparam to select which slot FreePIE will read:

freepie_slot = 0