Further progress for the dynamic cockpit.
All texture resolutions work for the dynamic cockpit now. The hi-res
textures load an additional alpha or glow texture that has to be
disabled. Also added another display that can be used to render
text.

The targeted object is now rendered as a shadeless texture (the
diffuse component is ignored). This allows the display to be
properly lit at all times.

Moved the code that loads/unloads the new textures to a different
location (DeviceResources OnSizeChanged), because that's where the
offscreen buffers are released/created. It may help with the crashes
I've seen in the past. If not, then the load/unload texture code
probably has to be moved somewhere else or disabled altogether.