Fixed the Dynamic Cockpit coordinate conversion in VR mode
The dynamic cockpit is now rendered using the original vertex
shader and original settings. That way, the in-game pixel coordinates
can be converted to dynamic cockpit RTV (screen-size) coordinates.
The formula looks like this:

screen_x = x / displayWidth * width + left

displayWidth is in-game resolution, like 1280, for instance.
width is the screen-size pixel size of the viewport. It's computed
by ddraw when the in-game resolution is scaled to screen res.
In my particular case, for in-game res of 1280x1024 and screen res
of 3240x2160, width is 2700 and height is 2160.

left, and top are the margins that are left over when the in-game
res doesn't fully cover the screen res. The viewport and constant
buffers must be set like this:

g_VSCBuffer.viewportScale[0] =  2.0f / displayWidth;
g_VSCBuffer.viewportScale[1] = -2.0f / displayHeight;
g_VSCBuffer.viewportScale[2] = scale;

viewport.TopLeftX = (float)left;
viewport.TopLeftY = (float)top;
viewport.Width = (float)width;
viewport.Height = (float)height;

This allows a predictable way to map in-game pixel coords to screen
coordinates.