HUD Overlay for Dynamic Cockpit is now working on the left image
The HUD overlay is now displayed on top of everything when the Dynamic
Cockpit is enabled. I still need to fix the right eye and the SteamVR
path. Also need to fix the background for the HUD (it is now transparent).
Stuff also needs to be erased from the overlay. Right now, everything is
duplicated.

Also added XWAObject.h and other files to read the state from XWA directly,
as if we were in a hook. This seems to work fine for now. For instance,
g_playerInHangar can be used to prevent the target arrow from being displayed.