Added ClearBox function to clear regions of the HUD offscreen buffer.
The logic needs to be fixed, though. We need to have a separate section
where we decide which regions will be cleared after the HUD offscreen
buffer has been used on the dynamic cockpit. Right now, it will clear
a region as soon as it's used and since regions overlap, that also clears
other stuff. For instance, the lasers clear the top of the targeting
computer.