Dynamic Cockpit elements can now be specified with coordinates.
The coordinates have the format:

src-Width,Height, src-x0,y0, src-x1,y1, dst-Width,Height, dst-x0,y0, dst-x1,y1

Where "source" is the offscreen HUD buffer in screen coordinates and "dst"
is the cockpit texture where the HUD element will be rendered. The coordinates
in dst are in pixels. For example, if the game is rendering at 3240x2160, then
this will capture the targeting computer:

dc_target_comp_uv_coords = 3240,2160, 1180,1720, 2060,2180,  256,128, 13,12, 242,115

Also got rid of the _dynCockpitAuxBuffer and _dynCockpitAuxSRV and their
associated DX calls -- not needed anymore.

The ClearHUD function is now broken, though. Some coordinate tranformation
must be done to re-enable it.