Textures are now named.
Using Jeremy's xwa_hook_textures_tag it's now possible to get a texture's
name, and a bunch of information with it. I'm including that hook and the
xwa_hook_time as well. The code was only slightly touched to keep the
texture name and the texture dump code has also been touched to save the
name when the texture is dumped.