Added erase_coords, textures for Y-Wing and A-wing and fixes.
Textures are now released in the OnSizeChanged event so that we don't
fill the memory with unused textures when switching crafts. At this point
the active DC textures are also disabled so that we start from a clean
state.

Added a bunch of textures for the Y-Wing and A-wing cockpits plus a sample
CFG file that contains coordinates for these cockpits as well as a fix for
the X-Wing so that the laser charge fills-up towards the center.