Added the HUD background for the Dynamic Cockpit
A new buffer that keeps the HUD background has been added. Shaders
have been modified to render the foreground and background of the
HUD. Some details still need to be fixed:

The targeted craft must be rendered over a "solid" background (it
looks semi-transparent right now)

Inactive regions are not cleared until their corresponding destination
textures are in view.