Fixed a bug in FreePIE. Dynamic Cockpit HUD works.
There was a nasty bug in FreePIE that disabled it entirely. It's been
fixed now. The Dynamic Cockpit now works along with the Fixed GUI
setting: it can either stick to your face or float inside the cockpit.

A new way to detect the cockpit was also added. Textures are tested for
"Cockpit" in their names. As soon as we detect one of those, we set the
proper cockpit. I think I added more regions too.