bgColors are uint, allowing 8 slots again.
Removed CRC code. bgColors for DC are now using uints instead of
float4's to save space and get 8 slots back. Tried to increase the
shader model to 5.0; but it keeps compiling to 4.0 level_9_3. Added
EnhanceLasers option to ddraw.cfg. Added flag to find light textures
and make them bright in the shader.